requirements.txt
flask>=3.0.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.0
flask-bcrypt>=1.0.1
apscheduler>=3.10.4
selenium>=4.18.0
cryptography>=42.0.0
python-dotenv>=1.0.0
waitress>=3.0.0
